Moved AbstractContentIntegrityCheck to an exported package
