+++ src/main/java/org/jahia/modules/verifyintegrity/services/checks/JCRLanguagePropertyCheck.java
@@ -3,10 +3,8 @@ package org.jahia.modules.verifyintegrity.services.checks;
 import org.apache.commons.lang.StringUtils;
 import org.jahia.api.Constants;
 import org.jahia.modules.verifyintegrity.api.ContentIntegrityCheck;
-import org.jahia.modules.verifyintegrity.services.AbstractContentIntegrityCheck;
+import org.jahia.modules.verifyintegrity.services.impl.AbstractContentIntegrityCheck;
 import org.jahia.modules.verifyintegrity.services.ContentIntegrityError;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
