+++ src/main/java/org/jahia/modules/verifyintegrity/services/checks/MarkForDeletionCheck.java
@@ -1,6 +1,6 @@
 package org.jahia.modules.verifyintegrity.services.checks;
 
-import org.jahia.modules.verifyintegrity.services.AbstractContentIntegrityCheck;
+import org.jahia.modules.verifyintegrity.services.impl.AbstractContentIntegrityCheck;
 import org.jahia.modules.verifyintegrity.services.ContentIntegrityError;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
