+++ src/main/java/org/jahia/modules/verifyintegrity/services/checks/PublicationSanityDefaultCheck.java
@@ -3,13 +3,11 @@ package org.jahia.modules.verifyintegrity.services.checks;
 import org.apache.commons.lang.StringUtils;
 import org.jahia.api.Constants;
 import org.jahia.modules.verifyintegrity.api.ContentIntegrityCheck;
-import org.jahia.modules.verifyintegrity.services.AbstractContentIntegrityCheck;
+import org.jahia.modules.verifyintegrity.services.impl.AbstractContentIntegrityCheck;
 import org.jahia.modules.verifyintegrity.services.ContentIntegrityError;
 import org.jahia.services.content.JCRNodeWrapper;
 import org.jahia.services.content.JCRSessionFactory;
 import org.jahia.services.content.JCRSessionWrapper;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
