+++ src/main/java/org/jahia/modules/verifyintegrity/services/checks/PublicationSanityLiveCheck.java
@@ -2,12 +2,10 @@ package org.jahia.modules.verifyintegrity.services.checks;
 
 import org.jahia.api.Constants;
 import org.jahia.modules.verifyintegrity.api.ContentIntegrityCheck;
-import org.jahia.modules.verifyintegrity.services.AbstractContentIntegrityCheck;
+import org.jahia.modules.verifyintegrity.services.impl.AbstractContentIntegrityCheck;
 import org.jahia.modules.verifyintegrity.services.ContentIntegrityError;
 import org.jahia.services.content.JCRSessionFactory;
 import org.jahia.services.content.JCRSessionWrapper;
-import org.osgi.service.component.ComponentContext;
-import org.osgi.service.component.annotations.Activate;
 import org.osgi.service.component.annotations.Component;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
